com.unity.xr.windowsmr.metro@1.0.1 # Changelog All notable changes to this package will be documented in this file.
The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).

## [1.0.1] - 2018-05-03

Update runtime .asmdef to remove Editor as it's not necessary

## [1.0.0] - 2018-04-20

First official version of the package.
Update name to com.unity.xr.windowsmr.metro

## [0.0.4] - 2018-04-04

Update Runtime assembly definition file to only support the specific platforms that WindowsMR scripts should work on.

## [0.0.3] - 2018-03-29

Update Assembly definitions for Editor and Runtime

## [0.0.2] - 2018-03-28

Update package name to com.unity.windowsmr.metro

## [0.0.1] - 2018-01-10

### This is the first release of *Unity Package Oculus*.

*Short description of this release*
This is the first test iteration of the Oculus XR Package using the Package Manager process. This package is a replica of the builds.zip hosted at:
https://ono.unity3d.com/unity-extra/vr-oculus-external#trunk

We are bringing over all libraries from the builds.zip to be included and distributed through the package manager.
This package contains the support for VR Device Oculus and for the Oculus Audio Spatializer.

The Current Oculus library version for this package is 1.21. Release notes for this version can be found here
(make sure to change the version in the drop on the page):
https://developer.oculus.com/downloads/package/oculus-utilities-for-unity-5/